fix: Type Index and other util files
